finished crud

+++ contacts/views.py
@@ -32,7 +32,7 @@ def create_contact(request):
 
     if request.method == 'POST':
         
-        form = ContactForm(data=request.POST)
+        form = ContactForm(request.POST, request.FILES)
 
         if form.is_valid():
             contact = form.save(commit=False)
@@ -41,22 +41,45 @@ def create_contact(request):
 
             if contact is not None:
                 return redirect('/contacts')
-    
+    else:
+        form = ContactForm()
+
     context = {'form': form, 'date': date}
 
     return render(request, 'contacts/create_contact.html', context)
 
 @login_required
-def edit_contact(request, id):
-    pass
+def edit_contact(request, pk):
+    
+    contact = Contact.objects.get(pk=pk)
+
+    form = ContactForm()
+
+    if request.method == 'POST':
+        form = ContactForm(request.POST, request.FILES, instance=contact)
+
+        if form.is_valid():
+            contact = form.save(commit=False)
+            contact.user = request.user
+            contact.save()
+
+            if contact is not None:
+                return redirect('/contacts')
+    else:
+        # with this we put the data that has the old contact
+        form = ContactForm(instance=contact)
+    
+    context = {'form': form, 'date': date}
+
+    return render(request, 'contacts/edit_contact.html', context)
 
 @login_required
-def delete_contact(request, id):
+def delete_contact(request, pk):
     
-    contact = Contact.objects.get(pk=id)
+    contact = Contact.objects.get(pk=pk)
 
-    # if contact is not None:
-    contact.delete()
+    if contact is not None:
+        contact.delete()
 
     # return render(request, 'contacts/list_contact.html', {})
     return redirect('/contacts')
